backend/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastapi-cli==0.0.5
flake8==7.1.1
flake8-bugbear==24.8.19
flake8-builtins==2.5.0
flake8-import-order==0.18.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
protobuf==5.27.3
protoc-exe==22.0rc1
psycopg2==2.9.9
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.2
pytest-subtests==0.13.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
redis==5.0.8
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.38.2
tomli==2.0.1
typer==0.12.4
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.23.0
websockets==13.0
